Generalize the boundary condition generation for clip_box
